Add snax_hwpe_mult as separate accelerator
